seperate the utlitiy functions from game_handler to utils module inside handler
