+++ game/handlers/game_handler.py
@@ -4,16 +4,22 @@ from game.handlers.database import (
     add_viewer,
     end_game_update,
     get_game_details,
-    get_game_users_sync,
     get_or_create_player,
     get_player_by_username,
     get_username,
-    get_viewer_details,
     update_game_players,
     update_game_state,
 )
-from game.handlers.error_handling import GameNotFoundError, PlayerNotFoundError, exception_handler
-from game.handlers.utils import create_message_data, notify, send_error
+from game.handlers.error_handling import exception_handler
+from game.handlers.utils import (
+    create_message_data,
+    determine_role,
+    game_users_list,
+    get_game_or_error,
+    get_game_users,
+    notify,
+    send_error,
+)
 
 
 async def route_event(consumer, data):
@@ -132,14 +138,6 @@ async def handle_game_get(consumer, data):
 
         await notify(consumer, consumer.channel_name, {'type': 'game.get.success', 'data': game_data})
 
-@exception_handler
-async def game_users_list(consumer, is_group=False):
-    """
-    Notify the list of users in a game.
-    """
-    users = await get_game_users(consumer.game_id)
-    await notify(consumer, consumer.room_group_name, {'type': 'game.users.list', 'data': users}, is_group=is_group)
-
 
 @exception_handler
 async def handle_game_end(consumer, data):
@@ -186,40 +184,3 @@ async def handle_game_end(consumer, data):
         {'type': 'game.end.success', 'message':  winning_username},
         is_group=True
     )
-
-async def get_game_users(game_id, viewer=None):
-    """
-    Asynchronously retrieve the list of users in a game, optionally including viewer details.
-    """
-    users = await get_game_users_sync(game_id)
-    if viewer:
-
-        viewer_details = await get_viewer_details(viewer)
-
-        return viewer_details
-
-    return users
-
-
-async def get_game_or_error(game_id):
-    game_data = await get_game_details(game_id)
-    if not game_data:
-
-        raise GameNotFoundError(f"Game ID {game_id} not found.")
-
-    return game_data
-
-async def get_player_or_error(username):
-    player = await get_player_by_username(username)
-    if not player:
-
-        raise PlayerNotFoundError(f"Player with username {username} not found.")
-
-    return player
-
-
-def determine_role(game_data):
-    """
-    Determine the role of a user in a game (player or viewer).
-    """
-    return 'viewer' if game_data['red_player'] and game_data['black_player'] else 'player'
